final WiotechStructureCreator with documentation
with additional documentation
